fix: remove spellcheck module
This module is deprecated and not used anymore
