--- data/taxonomies/fr_tokens_lower.gz
@@ -1,3 +1,0 @@
-version https://git-lfs.github.com/spec/v1
-oid sha256:41f18aeb07263134bb5228a7bd1a001b39767a0c4e4fd1f5512ef51fbb3ea85b
-size 3214511
