--- robotoff/elasticsearch/__init__.py
@@ -1,6 +1,0 @@
-"""
-This modules host Elasticsearch based methods to compute various insight
-"""
-
-# flake8: noqa
-from .client import get_es_client
