--- robotoff/elasticsearch/client.py
@@ -1,9 +1,0 @@
-import elasticsearch
-
-from robotoff import settings
-
-
-def get_es_client() -> elasticsearch.Elasticsearch:
-    return elasticsearch.Elasticsearch(
-        f"http://{settings.ELASTIC_USER}:{settings.ELASTIC_PASSWORD}@{settings.ELASTIC_HOST}:9200"
-    )
