--- robotoff/spellcheck/exceptions.py
@@ -1,10 +1,0 @@
-class SpellcheckerException(Exception):
-    pass
-
-
-class TokenLengthMismatchException(SpellcheckerException):
-    pass
-
-
-class LanguageNotAllowedException(SpellcheckerException):
-    pass
