+++ tests/unit/cli/test_main.py
@@ -5,10 +5,9 @@ def test_init_elasticsearch(mocker):
     fake_exporter = mocker.MagicMock()
 
     mocker.patch(
-        "robotoff.elasticsearch.export.ElasticsearchExporter",
+        "robotoff.elasticsearch.ElasticsearchExporter",
         return_value=fake_exporter,
     )
 
-    init_elasticsearch(True)
+    init_elasticsearch()
     fake_exporter.load_all_indices.assert_has_calls([])
-    fake_exporter.export_index_data.assert_has_calls([])
